createStacks --> studyMetadata.createDisplaySets(sopClassHandlerModules) (replaces `loadStudy`, `createDisplaySets` and `createStacks`)
