+++ src/classes/metadata/index.js
@@ -1,14 +1,26 @@
-import { StudyMetadata } from './StudyMetadata';
-import { SeriesMetadata } from './SeriesMetadata';
 import { InstanceMetadata } from './InstanceMetadata';
-import { OHIFStudyMetadata } from './OHIFStudyMetadata';
-import { OHIFSeriesMetadata } from './OHIFSeriesMetadata';
-import { OHIFInstanceMetadata } from './OHIFInstanceMetadata';
 import { Metadata } from './Metadata';
-import { WadoRsMetaDataBuilder } from './WadoRsMetaDataBuilder';
+import { OHIFInstanceMetadata } from './OHIFInstanceMetadata';
+import { OHIFSeriesMetadata } from './OHIFSeriesMetadata';
+import { OHIFStudyMetadata } from './OHIFStudyMetadata';
+import { SeriesMetadata } from './SeriesMetadata';
+import { StudyMetadata } from './StudyMetadata';
 import { StudySummary } from './StudySummary';
+import { WadoRsMetaDataBuilder } from './WadoRsMetaDataBuilder';
 
 const metadata = {
+  Metadata,
+  StudySummary,
+  WadoRsMetaDataBuilder,
+  StudyMetadata,
+  SeriesMetadata,
+  InstanceMetadata,
+  OHIFStudyMetadata,
+  OHIFSeriesMetadata,
+  OHIFInstanceMetadata,
+};
+
+export {
   Metadata,
   StudySummary,
   WadoRsMetaDataBuilder,
