+++ src/utils/index.js
@@ -1,14 +1,13 @@
+import { DICOMTagDescriptions } from './DICOMTagDescriptions';
 import ObjectPath from './objectPath';
-import guid from './guid';
+import StackManager from './StackManager.js';
 import absoluteUrl from './absoluteUrl';
 import addServers from './addServers';
+import guid from './guid';
 import sortBy from './sortBy.js';
-import writeScript from './writeScript.js';
-import StackManager from './StackManager.js';
-import createStacks from './createStacks.js';
-import createDisplaySets from './createDisplaySets.js';
 import studyMetadataManager from './studyMetadataManager';
-import { DICOMTagDescriptions } from './DICOMTagDescriptions';
+import updateMetaDataManager from './updateMetaDataManager.js';
+import writeScript from './writeScript.js';
 
 const utils = {
   guid,
@@ -18,9 +17,23 @@ const utils = {
   sortBy,
   writeScript,
   StackManager,
-  createStacks,
-  createDisplaySets,
   studyMetadataManager,
+  // Updates WADO-RS metaDataManager
+  updateMetaDataManager,
+  DICOMTagDescriptions,
+};
+
+export {
+  guid,
+  ObjectPath,
+  absoluteUrl,
+  addServers,
+  sortBy,
+  writeScript,
+  StackManager,
+  studyMetadataManager,
+  // Updates WADO-RS metaDataManager
+  updateMetaDataManager,
   DICOMTagDescriptions,
 };
